Womens March Shapefile
The content of this project itself is licensed under the Creative Commons Attribution Share Alike 4.0 license, and the underlying source code used to format and display that content is licensed under the MIT license. Work covered by the CC incorporates third party content that has been prominently marked and is not pursuant to this notice. 	

Thompson, Ashley C. "Land of the Free: A woman's place in critical geography" master's thesis in Risk & Resilience, Harvard University Graduate School of Design, 2017.